fix(wextra): fix -Wextra errors in debug_tools and device tests (round 6)
- EthernetConfig/DramConfig: add = {} defaults to compile_args, defines, named_compile_args
  (root-cause fix for all EthernetConfig/DramConfig designated-initializer warnings)
- ConfigRegPrintTestConfig: add = {} defaults to all fields (print_kernel etc)
- print_tile_helpers.hpp: fix int vs size_t loop vars
- test_print_tiles_multiple.cpp (dprint+device_print): fix int vs uint32_t
- test_print_output.cpp: fix int vs size_t in counts.size() comparison
- test_device.cpp: fix int vs size_t loop vars for grid_size/logical_grid_size

+++ tests/tt_metal/tt_metal/debug_tools/device_print/test_print_output.cpp
@@ -164,7 +164,7 @@ TEST_F(DevicePrintOutputFixture, PrintConcurrentAllRiscs) {
                 break;
             }
             int iter = -1;
-            if (sscanf(line.c_str(), "Test iteration: %d", &iter) == 1 && iter >= 0 && iter < counts.size()) {
+            if (sscanf(line.c_str(), "Test iteration: %d", &iter) == 1 && iter >= 0 && iter < static_cast<int>(counts.size())) {
                 counts[iter]++;
             }
         }
